ui/dataset: Extract dataset table cell into component file
